* added integration tests * fixed type bug in Issue

+++ src/main/java/de/linsin/github/rest/domain/Issue.java
@@ -64,7 +64,7 @@ public class Issue {
         }
     }
 
-    public long getNumber() {
+    public int getNumber() {
         return number;
     }
